+++ src/modules/accommodation/components/accommodation-details.tsx
