+++ src/modules/accommodation/components/amenities-modal.tsx
