+++ src/modules/accommodation/components/skeleton-details.tsx
