+++ src/modules/accommodation/components/skeleton-images.tsx
