+++ src/modules/accommodation/constants/icons.constant.ts
