+++ src/modules/accommodation/utils/get-amenities.util.ts
